// edudoc/src/app/model/interfaces/school-district.d.ts
export interface ISchoolDistrict extends IEntity {
    Name: string;
    Code: string;
    EinNumber: string;
    IrnNumber: string;
    NpiNumber: string;
    ProviderNumber: string;
    BecameTradingPartnerDate?: Date;
    BecameClientDate?: Date;
    RevalidationDate?: Date;
    ValidationExpirationDate?: Date;
    ProgressReports: boolean;
    ProgressReportsSent?: Date;
    RequireNotesForAllEncountersSent: boolean;
    NotesRequiredDate?: Date;
    Notes?: string;
    CreatedById: number;
    ModifiedById?: number;
    DateCreated?: Date;
    DateModified?: Date;
    UseDisabilityCodes: boolean;
    Archived: boolean;
    AddressId?: number;
    AccountManagerId?: number;
    AccountAssistantId?: number;
    TreasurerId?: number;
    SpecialEducationDirectorId?: number;
    MerId?: number;
    ActiveStatus: boolean;
    CaseNotesRequired: boolean;
    IepDatesRequired: boolean;

    // reverse nav
    ActivitySummaryDistricts?: IActivitySummaryDistrict[];
    AdminSchoolDistricts?: IAdminSchoolDistrict[];
    AnnualEntries?: IAnnualEntry[];
    BillingScheduleDistricts?: IBillingScheduleDistrict[];
    ClaimsDistricts?: IClaimsDistrict[];
    Contacts?: IContact[];
    DistrictProgressReportDates?: IDistrictProgressReportDate[];
    EncounterIdentifiers?: IEncounterIdentifier[];
    EscSchoolDistricts?: IEscSchoolDistrict[];
    Messages?: IMessage[];
    MessageDocuments?: IMessageDocument[];
    MessageLinks?: IMessageLink[];
    PendingReferrals?: IPendingReferral[];
    ProviderCaseUploads?: IProviderCaseUpload[];
    ProviderCaseUploadDocuments?: IProviderCaseUploadDocument[];
    ProviderEscSchoolDistricts?: IProviderEscSchoolDistrict[];
    RosterValidationDistricts?: IRosterValidationDistrict[];
    SchoolDistrictProviderCaseNotes?: ISchoolDistrictProviderCaseNote[];
    SchoolDistrictRosters?: ISchoolDistrictRoster[];
    SchoolDistrictRosterDocuments?: ISchoolDistrictRosterDocument[];
    SchoolDistrictsAccountAssistants?: ISchoolDistrictsAccountAssistant[];
    SchoolDistrictsFinancialReps?: ISchoolDistrictsFinancialRep[];
    SchoolDistrictsSchools?: ISchoolDistrictsSchool[];
    Students?: IStudent[];
    StudentDistrictWithdrawals?: IStudentDistrictWithdrawal[];
    UnmatchedClaimRespons?: IUnmatchedClaimRespons[];
    Users_DistrictAdminId?: IUser[];
    Users_SchoolDistrictId?: IUser[];
    Vouchers?: IVoucher[];

    // foreign keys
    Address?: IAddress;
    SpecialEducationDirector?: IContact;
    Treasurer?: IContact;
    MerFile?: IDocument;
    AccountAssistant?: IUser;
    AccountManager?: IUser;
    CreatedBy?: IUser;
    ModifiedBy?: IUser;
}